server/requirements.txt
fastapi
uvicorn
pandas
holoviews
hvplot
panel
bokeh_fastapi
